nanobot-dashboard/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.20
croniter>=6.0.0
python-frontmatter>=1.1.0
pydantic>=2.0.0
aiofiles>=24.0.0
